requirements.txt
python-telegram-bot==20.3
openai==1.14.2
supabase==2.2.1
python-dotenv==1.0.1